chore(macos): align Sky Garden extension with current runtime
