+++ macos/README.md
@@ -10,11 +10,21 @@ This project injects through **local loopback CDP**. It does **not** modify the 
 
 ## Requirements
 
-- macOS
+- macOS 13 Ventura or newer (the native DMG app declares macOS 13 as its minimum)
 - Official Codex Desktop installed and launched at least once (`~/.codex/config.toml` exists)
 - No global Node.js install required (uses Codex’s signed bundled Node after validation)
 
-## Quick start (from this repo)
+## Release install (recommended)
+
+普通用户请从 [GitHub Releases](https://github.com/Fei-Away/Codex-Dream-Skin/releases) 下载
+`CodexDreamSkin-vX.Y.Z.dmg`，按 [`docs/install-macos.md`](../docs/install-macos.md) 的图形界面步骤
+拖入 Applications。首次运行可能需要在“系统设置 → 隐私与安全性 → 仍要打开”确认一次；不需要
+运行 `xattr` 或安装源码。后续更新下载新的 DMG 覆盖安装即可，用户主题和图片会保留。
+
+## Advanced: run from source
+
+The Release DMG above is the normal user path. The commands below are for
+contributors, diagnostics, and legacy deployments.
 
 ```bash
 # 1) Optional checks (needs the installed Codex/ChatGPT.app bundled Node)
@@ -33,7 +43,7 @@ This project injects through **local loopback CDP**. It does **not** modify the 
 #    Codex Dream Skin - Verify.command
 #    Codex Dream Skin - Restore.command
 
-# 5) Optional: menu bar (SwiftBar) — apply, pause, import, and switch
+# 5) Legacy only: install the old SwiftBar menu (do not enable it beside the native app)
 ./Install\ Menu\ Bar.command
 # Look for 🎨 Skin in the top-right menu bar
 ```
@@ -47,7 +57,7 @@ Install location after step 2:
 | Theme backup | under Application Support (`theme-backup.json`) |
 | Bundled Codex pet | `~/.codex/pets/dream-skin-sky-garden-duo` |
 
-## Customer ZIP (optional packaging)
+## Legacy standalone ZIP (maintainer/offline packaging only)
 
 To build the “double-click install” folder layout for non-git users:
 
@@ -55,7 +65,12 @@ To build the “double-click install” folder layout for non-git users:
 ./scripts/build-client-release.sh "$HOME/Desktop/Codex 主题编辑器.zip"
 ```
 
-That ZIP contains a visible installer plus a hidden `.codex-dream-skin-studio` engine. Do not ship only CSS/images.
+That ZIP contains a visible installer plus a hidden `.codex-dream-skin-studio`
+engine and is staged as a rights-clean package with only the redistributable
+Gothic Void Crusade preset. It is retained for existing offline workflows;
+prefer the DMG for ordinary users, and do not share a source checkout or an
+archive containing the excluded Arina reference files. Do not ship only
+CSS/images.
 
 ## How it works (security boundary)
 
@@ -75,13 +90,16 @@ CDP is powerful and unauthenticated on loopback. Prefer Restore when you are don
 
 ## Bundled presets
 
-A fresh install seeds three tested presets into your theme library:
-**Gothic Void Crusade**, **桥本有菜 / Arina Hashimoto**, and
-**天空花园双姝 / Sky Garden Duo**. Gothic Void Crusade is the default when no
-active theme exists. Switch themes with:
+The public macOS package seeds two redistributable presets:
+**Gothic Void Crusade** and **天空花园双姝 / Sky Garden Duo**. Gothic Void
+Crusade remains the default when no active theme exists. A source checkout also
+contains the **桥本有菜 / Arina Hashimoto** reference material, but the public
+app bundle deliberately excludes it until independent likeness and
+redistribution rights are confirmed.
+
+Switch to Sky Garden Duo with:
 
 ```bash
-~/.codex/codex-dream-skin-studio/scripts/switch-theme-macos.sh --id preset-arina-hashimoto
 ~/.codex/codex-dream-skin-studio/scripts/switch-theme-macos.sh --id preset-sky-garden-duo
 ```
 
@@ -106,7 +124,8 @@ before redistributing it.
 
 Seeding is idempotent. Upgrades remove only retired bundled preset IDs; your
 own `custom-*` themes from “换一张图” and the currently active theme copy are
-never touched.
+never touched. Existing locally saved reference themes are not deleted by an
+upgrade, but they are not copied into newly downloaded public packages.
 
 To contribute a preset, see [`presets/README.md`](./presets/README.md).
